jira-mcp-app/pyproject.toml
[project]
name = "jira-mcp"
version = "0.1.0"
description = "Jira Project Management MCP server for The Great Trading Company"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]==1.26.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "uvicorn>=0.29",
    "starlette>=0.37",
    "structlog>=24.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
jira-mcp = "jira_mcp.server:main"
